requirements.txt
flask
tensorflow
numpy
pillow
opencv-python
imutils
tqdm
scikit-learn
gunicorn